feat viewchange

+++ Assets/3.Script/Interface/IDamage.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public interface IDamage
 {
     //피해량,맞은위치,맞은각도
-    void OnDamage(float damage, Vector3 hitPosition, Vector3 hitNomal);
+    void OnDamage(int damage, Vector3 hitPosition, Vector3 hitNomal);
 }
